+++ hydra/blocks/spinhalf/spinhalf.h
@@ -16,18 +16,18 @@ public:
   using indexing_t = indexing::SpinhalfIndexing;
 
   Spinhalf() = default;
-  Spinhalf(int n_sites);
-  Spinhalf(int n_sites, int n_up);
-  Spinhalf(int n_sites, PermutationGroup permutation_group,
-           Representation irrep, int n_sublat = 0);
-  Spinhalf(int n_sites, int n_up, PermutationGroup permutation_group,
-           Representation irrep, int n_sublat = 0);
+  Spinhalf(int64_t n_sites);
+  Spinhalf(int64_t n_sites, int64_t n_up);
+  Spinhalf(int64_t n_sites, PermutationGroup permutation_group,
+           Representation irrep, int64_t n_sublat = 0);
+  Spinhalf(int64_t n_sites, int64_t n_up, PermutationGroup permutation_group,
+           Representation irrep, int64_t n_sublat = 0);
 
-  inline int n_sites() const { return n_sites_; }
+  inline int64_t n_sites() const { return n_sites_; }
   inline bool sz_conserved() const { return sz_conserved_; }
-  inline int sz() const { return sz_; }
-  inline int n_up() const { return n_up_; }
-  inline int n_dn() const { return n_dn_; }
+  inline int64_t sz() const { return sz_; }
+  inline int64_t n_up() const { return n_up_; }
+  inline int64_t n_dn() const { return n_dn_; }
   inline bool symmetric() const { return symmetric_; }
   inline PermutationGroup const &permutation_group() const {
     return permutation_group_;
@@ -41,13 +41,13 @@ public:
   indexing_t const &indexing() const;
 
 private:
-  int n_sites_;
+  int64_t n_sites_;
   bool sz_conserved_;
-  int n_up_;
-  int n_dn_;
-  int sz_;
+  int64_t n_up_;
+  int64_t n_dn_;
+  int64_t sz_;
   bool symmetric_;
-  int n_sublat_;
+  int64_t n_sublat_;
   PermutationGroup permutation_group_;
   Representation irrep_;
   std::shared_ptr<indexing_t> indexing_;
